Moved tests on scanning to a separate module, test_scanning.
